refactor(main): extract panel routing into host/panelView
