+++ src/main/index.ts
@@ -1,5 +1,5 @@
-import { app, Menu, ipcMain, net, WebContentsView } from 'electron'
-import type { BrowserWindow } from 'electron'
+import { app, Menu, ipcMain, net } from 'electron'
+import type { BrowserWindow, WebContentsView } from 'electron'
 // Type-only while docking-to-tray is disabled.
 import type { Tray } from 'electron'
 import path from 'path'
@@ -37,8 +37,8 @@ import { registerAssetDownloadHandlers } from './lib/ipc/registerAssetDownloadHa
 import { get as getInstallation, installationEvents } from './installations'
 import { showModelFolderRelaunchPage } from './lib/relaunchPage'
 import { COMFY_BG, SPLASH_DARK, TITLEBAR_BG, type SplashTheme } from './lib/theme'
-import { TITLEBAR_HEIGHT, comfyTitleBarOverlay } from './lib/titleBarOverlay'
-import { resolveTheme, sourceMap, _registerExtraBroadcastTarget, _broadcastToRenderer } from './lib/ipc/shared'
+import { comfyTitleBarOverlay } from './lib/titleBarOverlay'
+import { sourceMap, _broadcastToRenderer } from './lib/ipc/shared'
 import * as mainTelemetry from './lib/telemetry'
 import { getDeviceId } from './lib/deviceId'
 
@@ -51,9 +51,7 @@ import {
   openOrFocusChooserHostWindow,
   pendingAttachClaims,
   setHostFactories,
-  VALID_PANELS,
 } from './host/registry'
-import type { BodyMode, ComfyPanelKey, ComfyWindowEntry } from './host/registry'
 import {
   applyChooserHostThemeToAll,
   createHostWindow,
@@ -71,6 +69,13 @@ import {
   returnToDashboard,
   setDetachFactories,
 } from './host/detach'
+import {
+  ensurePanelView,
+  refreshComfyTabBody,
+  registerPanelViewIpc,
+  sendToPanelDeferred,
+  setActivePanel,
+} from './host/panelView'
 
 export type { ComfyPanelKey } from './host/registry'
 
@@ -81,29 +86,6 @@ const APP_VERSION = getAppVersion()
 // The chooser host window plus per-install ComfyUI windows are the
 // only top-level surfaces.
 let tray: Tray | null = null
-
-/**
- * Re-evaluate the body mode for a comfy window after a session-state
- * transition (instance launched / stopped / crashed) and reflect it in the
- * layout. When the body mode is `'comfy-lifecycle'`, the panelView is created
- * (if needed) and asked to render the lifecycle UI; the title-bar pill stays
- * on `'comfy'` either way.
- */
-function refreshComfyTabBody(installationId: string): void {
-  const entry = getEntryByInstallationId(installationId)
-  if (!entry || entry.window.isDestroyed()) return
-  if (entry.activePanel !== 'comfy') return
-
-  const mode = computeBodyMode(entry)
-  if (mode === 'comfy-lifecycle') {
-    const panelView = ensurePanelView(entry.windowKey, entry, 'comfy-lifecycle')
-    if (!panelView.webContents.isDestroyed() && !panelView.webContents.isLoadingMainFrame()) {
-      panelView.webContents.send('panel-switch', { panel: 'comfy-lifecycle', installationId })
-    }
-  }
-  entry.layoutViews()
-  focusActiveBody(entry)
-}
 
 function focusExternalProcessWindow(pid: number): void {
   if (process.platform === 'win32') {
@@ -441,174 +423,6 @@ ipcMain.handle('quit-app', () => quitApp())
 // straggling renderer still bound to the channel doesn't reject.
 ipcMain.handle('reset-zoom', () => {
   // no-op
-})
-
-/**
- * Lazily create the panel WebContentsView for a comfy window. Adds it as a
- * sibling of comfyView, registers it for broadcasts, and loads panel.html
- * with the installation context as URL parameters.
- *
- * The URL params are only an initial hint — `did-finish-load` always re-pushes
- * the current `activePanel` so a user who clicks Install Settings then
- * Launcher Settings before the first load completes still ends up on the
- * latter. This guards against the mid-load race.
- */
-function ensurePanelView(windowKey: number, entry: ComfyWindowEntry, initialPanel: BodyMode): WebContentsView {
-  if (entry.panelView) return entry.panelView
-
-  const panelView = new WebContentsView({
-    webPreferences: {
-      nodeIntegration: false,
-      contextIsolation: true,
-      // sandbox: false — preload/index.js imports the shared
-      // src/preload/api.ts chunk, same as the title-bar preload.
-      // Sandboxed preloads can't require() relative chunks, so leaving
-      // sandbox on would silently break window.api in the panel and
-      // take renderer-side telemetry with it. See issue #521 for the
-      // build-time inlining plugin that will let us turn sandbox back on.
-      sandbox: false,
-      // Reuse the launcher preload — panel UI uses window.api like the main launcher window.
-      preload: path.join(__dirname, '../preload/index.js'),
-      // Default session (no partition) — keeps the panel isolated from the
-      // ComfyUI frontend's storage even though it runs in the same window.
-    },
-  })
-  panelView.setBackgroundColor(resolveTheme() === 'dark' ? '#202020' : '#ffffff')
-  entry.window.contentView.addChildView(panelView)
-  // Insert at zero size, behind the comfy view; layoutViews handles positioning.
-  panelView.setBounds({ x: 0, y: TITLEBAR_HEIGHT + 1, width: 0, height: 0 })
-  panelView.setVisible(false)
-
-  // Push the *latest* body mode (may differ from initialPanel if the user
-  // clicked between buttons during the first load, or the running state
-  // changed) and steal focus if the window is focused.
-  panelView.webContents.once('did-finish-load', () => {
-    const latest = comfyWindows.get(windowKey)
-    if (!latest || latest.window.isDestroyed() || panelView.webContents.isDestroyed()) return
-    const mode = computeBodyMode(latest)
-    if (mode !== 'comfy') {
-      panelView.webContents.send('panel-switch', { panel: mode, installationId: latest.installationId ?? '' })
-      if (latest.window.isFocused()) panelView.webContents.focus()
-    }
-  })
-
-  // Pass the entry's installationId (which is the empty string for
-  // install-less host windows) to the panel renderer — the Map key is a
-  // numeric windowKey that PanelApp.vue must not see.
-  const panelInstallationId = entry.installationId ?? ''
-  const isDev = !!process.env['ELECTRON_RENDERER_URL']
-  const loadPromise = isDev
-    ? panelView.webContents.loadURL(
-        `${(process.env['ELECTRON_RENDERER_URL'] as string).replace(/\/$/, '')}/panel.html?installationId=${encodeURIComponent(panelInstallationId)}&panel=${encodeURIComponent(initialPanel)}`,
-      )
-    : panelView.webContents.loadFile(
-        path.join(__dirname, '../renderer/panel.html'),
-        { query: { installationId: panelInstallationId, panel: initialPanel } },
-      )
-  // Loads can reject if the window closes mid-load — swallow to avoid noisy
-  // unhandledRejection forwarding from the main-process error handlers.
-  void loadPromise.catch(() => {})
-
-  _registerExtraBroadcastTarget(panelView.webContents)
-  entry.panelView = panelView
-  return panelView
-}
-
-/** Move OS focus to whichever body view is now active so keyboard input lands in the right place. */
-function focusActiveBody(entry: ComfyWindowEntry): void {
-  if (entry.window.isDestroyed() || !entry.window.isFocused()) return
-  const mode = computeBodyMode(entry)
-  if (mode === 'comfy') {
-    if (!entry.comfyView.webContents.isDestroyed()) entry.comfyView.webContents.focus()
-  } else if (entry.panelView && !entry.panelView.webContents.isDestroyed() && !entry.panelView.webContents.isLoadingMainFrame()) {
-    // Panel exists and is loaded — focus immediately. If still loading, the
-    // did-finish-load handler in ensurePanelView will focus it.
-    entry.panelView.webContents.focus()
-  }
-}
-
-function setActivePanel(windowKey: number, panel: ComfyPanelKey): void {
-  const entry = comfyWindows.get(windowKey)
-  if (!entry || entry.window.isDestroyed()) return
-  // The unified Settings modal works in both install-backed and install-
-  // less hosts (PanelApp picks the appropriate default tab — ComfyUI
-  // Settings vs Global Settings — at mount time), so no install-less
-  // gating is required here.
-
-  if (entry.activePanel === panel) return
-
-  entry.activePanel = panel
-  // Resolve to the actual body mode (Comfy pill maps to lifecycle / chooser
-  // depending on running state and whether the window is install-backed).
-  const mode = computeBodyMode(entry)
-  if (mode !== 'comfy') {
-    const panelView = ensurePanelView(windowKey, entry, mode)
-    // If panel view already loaded, push the switch immediately. If still
-    // loading, the did-finish-load handler in ensurePanelView will push the
-    // current body mode — guarding against rapid clicks during first load.
-    if (!panelView.webContents.isDestroyed() && !panelView.webContents.isLoadingMainFrame()) {
-      panelView.webContents.send('panel-switch', { panel: mode, installationId: entry.installationId ?? '' })
-    }
-  }
-  entry.layoutViews()
-  if (!entry.titleBarView.webContents.isDestroyed()) {
-    // Title bar pill stays on the user-visible key — never reflects the
-    // internal `'comfy-lifecycle'` body mode.
-    entry.titleBarView.webContents.send('comfy-titlebar:panel-changed', panel)
-  }
-  focusActiveBody(entry)
-}
-
-ipcMain.on('comfy-window:set-panel', (event, payload: { panel: string }) => {
-  const found = findEntryByTitleBarSender(event.sender)
-  if (!found) return
-  const panel = payload?.panel as ComfyPanelKey
-  if (!VALID_PANELS.has(panel)) return
-  setActivePanel(found.id, panel)
-})
-
-/**
- * Send a payload to a panelView, deferring until `did-finish-load` if
- * the bundle is still loading.
- *
- * Title-bar pill clicks and popup deep-links can land while the panel
- * renderer is still booting — the panelView is constructed lazily on
- * the first non-comfy switch, so its preload + Vue app aren't ready
- * yet on the very first click. A synchronous `send()` then arrives
- * before the renderer's `onPanelTriggerOverlay` (or other) listener
- * runs in `onMounted`, and the IPC is silently dropped. This helper
- * centralizes the deferral pattern used by every such handler.
- */
-function sendToPanelDeferred(panelView: WebContentsView, channel: string, payload: unknown): void {
-  if (panelView.webContents.isDestroyed()) return
-  const send = (): void => {
-    if (panelView.webContents.isDestroyed()) return
-    panelView.webContents.send(channel, payload)
-  }
-  if (panelView.webContents.isLoadingMainFrame()) {
-    panelView.webContents.once('did-finish-load', send)
-  } else {
-    send()
-  }
-}
-
-/**
- * Page-level X close (rendered inside the panel WebContentsView, e.g.
- * Settings / Directories / Install Settings) — same effect as a pill
- * click: the body returns to the comfy/chooser root. The panel preload
- * exposes this as `closeCurrentPanel()`.
- *
- * We resolve the host window via the panel's WebContents sender. The
- * panelView is lazily created so we walk every entry instead of caching
- * a separate reverse-map.
- */
-ipcMain.on('comfy-window:close-current-panel', (event) => {
-  for (const [id, entry] of comfyWindows) {
-    if (entry.panelView?.webContents === event.sender) {
-      setActivePanel(id, 'comfy')
-      return
-    }
-  }
 })
 
 /**
@@ -1027,6 +841,7 @@ if (app.isPackaged && !app.requestSingleInstanceLock()) {
     })
     setDetachFactories({ ensurePanelView })
     setHostFactories({ createChooser: openChooserHostWindow })
+    registerPanelViewIpc()
 
     migrateXdgPaths()
     registerProcessErrorHandlers()
